Build LavaMoat bundles concurrently
