+++ packages/snaps-execution-environments/package.json
@@ -28,8 +28,9 @@
     "build:tsc": "tsc --project tsconfig.build.json",
     "build": "yarn build:tsc && yarn build:post-tsc",
     "build:test": "yarn build:lavamoat",
-    "build:lavamoat": "lavamoat scripts/build.js --policy lavamoat/build-system/policy.json  --policyOverride lavamoat/build-system/policy-override.json",
-    "build:lavamoat:policy": "yarn build:lavamoat --writeAutoPolicy && node scripts/build.js --writeAutoPolicy",
+    "build:lavamoat:single": "yarn lavamoat scripts/build.js --policy lavamoat/build-system/policy.json  --policyOverride lavamoat/build-system/policy-override.json",
+    "build:lavamoat": "node scripts/build-all.js",
+    "build:lavamoat:policy": "yarn build:lavamoat:single --writeAutoPolicy && yarn build:lavamoat --writeAutoPolicy",
     "auto-changelog-init": "auto-changelog init",
     "prepare-manifest:preview": "../../scripts/prepare-preview-manifest.sh",
     "publish:preview": "yarn npm publish --tag preview"
@@ -78,6 +79,7 @@
     "browserify": "^17.0.0",
     "buffer": "^6.0.3",
     "common-shakeify": "^1.1.2",
+    "concurrently": "^8.0.1",
     "deepmerge": "^4.2.2",
     "esbuild": "^0.17.15",
     "eslint": "^8.27.0",
